+++ src/main/groovy/all/shared/gradle/quality/code/BaseStyleConfigWrapper.groovy
@@ -2,11 +2,10 @@
 //  Licensed under the MIT License (MIT), see LICENSE.txt
 package all.shared.gradle.quality.code
 
-import all.shared.gradle.quality.code.config.CodeStyleConfig
-
 import groovy.transform.CompileStatic
 
 import org.gradle.api.Project
+import org.gradle.api.artifacts.Configuration
 import org.gradle.api.artifacts.repositories.MavenArtifactRepository
 
 @CompileStatic
@@ -17,17 +16,19 @@ class BaseStyleConfigWrapper {
 
   private BaseStyleConfigWrapper() { }
 
-  static void addExtension(final Project project) {
-    if (project.extensions.findByName(CodeStyleConfig.CODE_STYLE_EXTENSION) == null) {
+  static boolean addExtension(final Project project, final String extensionName) {
+    if (project.extensions.findByName(extensionName) == null) {
+      final Configuration configuration = project.configurations.maybeCreate(extensionName)
       project.repositories.maven { MavenArtifactRepository mavenRepo -> mavenRepo.url = QUALITY_CONFIG_MAVEN_REPO_URL }
-      project.configurations.maybeCreate(CodeStyleConfig.CODE_STYLE_EXTENSION)
-      project.dependencies.add(CodeStyleConfig.CODE_STYLE_EXTENSION,
+      project.dependencies.add(extensionName,
         "$QUALITY_CONFIG_MAVEN_COORDINATES:${project.findProperty(BASE_STYLE_CONFIG_VERSION_PROPERTY) ?: '+'}")
-      project.extensions.add(CodeStyleConfig.CODE_STYLE_EXTENSION, CodeStyleConfig.of(project))
-      project.logger.debug('Added code-style-check extension')
+      project.extensions.add(extensionName, BaseStyleConfigWrapperExtension.of(project.resources.text, configuration))
+      project.logger.debug('Added base-style-config-wrapper extension')
+      true
     }
     else {
-      project.logger.error('Couldn\'t add code-style-check extension')
+      project.logger.error('Couldn\'t add base-style-config-wrapper extension')
+      false
     }
   }
 }