+++ src/main/groovy/all/shared/gradle/quality/code/BaseStyleConfigWrapperExtension.groovy
@@ -1,22 +1,23 @@
 //  Copyright (c) 2018 Gonzalo Müller Bravo.
 //  Licensed under the MIT License (MIT), see LICENSE.txt
-package all.shared.gradle.quality.code.config
+package all.shared.gradle.quality.code
+
+import all.shared.gradle.quality.code.config.BackCodeStyleConfig
+import all.shared.gradle.quality.code.config.CommonCodeStyleConfig
+import all.shared.gradle.quality.code.config.FrontCodeStyleConfig
 
 import groovy.transform.CompileStatic
 
-import org.gradle.api.Project
 import org.gradle.api.artifacts.Configuration
 import org.gradle.api.resources.TextResourceFactory
 
 @CompileStatic
-class CodeStyleConfig {
-  public static final String CODE_STYLE_EXTENSION = 'baseStyleConfig'
-
+class BaseStyleConfigWrapperExtension {
   private final BackCodeStyleConfig back
   private final CommonCodeStyleConfig common
   private final FrontCodeStyleConfig front
 
-  private CodeStyleConfig(
+  private BaseStyleConfigWrapperExtension(
       final CommonCodeStyleConfig common,
       final BackCodeStyleConfig back,
       final FrontCodeStyleConfig front) {
@@ -25,10 +26,10 @@ class CodeStyleConfig {
     this.front = front
   }
 
-  static final CodeStyleConfig of(final Project project) {
-    final TextResourceFactory factory = project.resources.text
-    final Configuration configuration = project.configurations.getByName(CODE_STYLE_EXTENSION)
-    new CodeStyleConfig(
+  static final BaseStyleConfigWrapperExtension of(
+      final TextResourceFactory factory,
+      final Configuration configuration) {
+    new BaseStyleConfigWrapperExtension(
       new CommonCodeStyleConfig(
         factory.fromArchiveEntry(configuration, 'common/common-checks.xml'),
         factory.fromArchiveEntry(configuration, 'gradle/gradle-rules.groovy')),