+++ CHANGELOG.md
@@ -1,1 +1,9 @@
 # Base Style Config Wrapper Change Log
+
+
+## 1.0.1 - October 2018
+
+* Adds two new static fields to the `BaseStyleConfigWrapperPlugin`.
+* Renames `CodeStyleConfig` to `BaseStyleConfigWrapperExtension`.
+* Fixes some log information
+* Updates README file.